move verifier to common
